Added BRK Syscall, added tests for it, added NQueens test with board as dynamic array using malloc
